Ignore the previous files. THIS IS THE RIGHT ONE!
I made a mistake when I set up the GUI. The way I did it before all
classes were independent and we won't be able to connect them together.
I figured it out how to do it, but I had to redo the whole thing. This
only contains the LogIn connected to the PatientHome. It shouldn't take
me too long to get back on track tomorrow.